requirements.txt
networkx
rtree
geopandas
scikit-learn
python-igraph
